Extract circular imports and register signals with an AppConfig.
